backend/rest_api/pyproject.toml
[project]
name = "rest-api"
version = "0.1.0"
description = ""
authors = [
    {name = "NandanaSanjith",email = "nandanasanjith2175@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi[standard] (>=0.116.2,<0.117.0)",
    "uvicorn[standard] (>=0.35.0,<0.36.0)",
    "pymongo (>=4.15.1,<5.0.0)",
    "stripe (>=13.0.1,<14.0.0)",
    "python-dotenv (>=1.1.1,<2.0.0)"
]

[tool.poetry]
packages = [{include = "rest_api", from = "src"}]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
